Internal library

Index

EMGExt.compliance_area_availability

EMGExt.compliance_bidirectional

EMGExt.compliance_con_rate

EMGExt.compliance_loss

EnergyModelsCompliance.compliance_capacity

EnergyModelsCompliance.compliance_capacity

EnergyModelsCompliance.compliance_data

EnergyModelsCompliance.compliance_data

EnergyModelsCompliance.compliance_inputs

EnergyModelsCompliance.compliance_inputs

EnergyModelsCompliance.compliance_level

EnergyModelsCompliance.compliance_opex_fixed

EnergyModelsCompliance.compliance_opex_fixed

EnergyModelsCompliance.compliance_opex_var

EnergyModelsCompliance.compliance_opex_var

EnergyModelsCompliance.compliance_outputs

EnergyModelsCompliance.compliance_outputs

EnergyModelsCompliance.compliance_stor_res

Nodes and Links

EnergyModelsCompliance.compliance_capacity — Method

compliance_capacity(n::EMB.Node)

Returns a Bool indicating whether the function capacity is applicable for the developed Node.

The function checks both capacity(n) and capacity(n, t) even if the former is not directly used.

EnergyModelsCompliance.compliance_opex_var — Method

compliance_opex_var(n::EMB.Node)
compliance_opex_var(n::Sink)

Returns a Bool indicating whether the function opex_var is applicable for the developed Node or alternatively the functions surplus_penalty and deficit_penalty.

The function checks both opex_var(n) (surplus_penalty(n) and deficit_penalty(n)) and opex_var(n, t) (surplus_penalty(n, t) and deficit_penalty(n, t)) even if the former is not directly used.

EnergyModelsCompliance.compliance_opex_fixed — Method

compliance_opex_fixed(n::EMB.Node)
compliance_opex_fixed(n::Sink)

Returns a Bool indicating whether the function opex_fixed is applicable for the developed Node.

The function checks both opex_fixed(n) and opex_fixed(n, t_inv) even if the former is not directly used.

EnergyModelsCompliance.compliance_inputs — Method

compliance_inputs(n::EMB.Node)

Returns a Tuple{Bool} indicating whether the function inputs is applicable for the developed Node.

The first entry of the Tuple is indicating whether the model can build while the second entry indicates whether problems with existing base functions may occur.

The function checks both inputs(n) and inputs(n, p).

EnergyModelsCompliance.compliance_outputs — Method

compliance_outputs(n::EMB.Node)

Returns a Tuple{Bool} indicating whether the function outputs is applicable for the developed Node.

The first entry of the Tuple is indicating whether the model can build while the second entry indicates whether problems with existing base functions may occur.

The function checks both outputs(n) and outputs(n, p).

EnergyModelsCompliance.compliance_data — Method

compliance_data(n::EMB.Node)

Returns a Bool indicating whether the function node_data is applicable for the developed Node.

EnergyModelsCompliance.compliance_level — Function

compliance_level(n::Storage)

Returns a Bool indicating whether the function level is applicable for the developed Storage.

EnergyModelsCompliance.compliance_stor_res — Function

compliance_stor_res(n::Storage)

Returns a Bool indicating whether the function storage_resource is applicable for the developed Storage.

Areas and TransmissionModes

EMGExt.compliance_area_availability — Function

compliance_area_availability(a::Area)

Returns a Bool indicating whether the function availability_node is applicable for the developed Area.

EnergyModelsCompliance.compliance_capacity — Method

EMC.compliance_capacity(tm::TransmissionMode)

Returns a Bool indicating whether the function capacity is applicable for the developed Node.

The function checks both capacity(tm) and capacity(tm, t) even if the former is not directly used.

EnergyModelsCompliance.compliance_opex_var — Method

EMC.compliance_opex_var(tm::TransmissionMode)

Returns a Bool indicating whether the function opex_var is applicable for the developed TransmissionMode.

The function checks both opex_var(tm) and opex_var(tm, t) even if the former is not directly used.

EnergyModelsCompliance.compliance_opex_fixed — Method

EMC.compliance_opex_fixed(tm::TransmissionMode)

Returns a Bool indicating whether the function opex_fixed is applicable for the developed TransmissionMode.

The function checks both opex_fixed(tm) and opex_fixed(tm, t_inv) even if the former is not directly used.

EnergyModelsCompliance.compliance_inputs — Method

EMC.compliance_inputs(tm::TransmissionMode)

Returns a Bool indicating whether the function inputs is applicable for the developed TransmissionMode.

The first entry of the Tuple is indicating whether the model can build while the second entry is always false.

EnergyModelsCompliance.compliance_outputs — Method

EMC.compliance_outputs(tm::TransmissionMode)

Returns a Bool indicating whether the function outputs is applicable for the developed TransmissionMode.

The first entry of the Tuple is indicating whether the model can build while the second entry is always false.

EnergyModelsCompliance.compliance_data — Method

EMC.compliance_data(tm::TransmissionMode)

Returns a Bool indicating whether the function mode_data is applicable for the developed TransmissionMode.

EMGExt.compliance_loss — Function

compliance_loss(tm::TransmissionMode)

Returns a Bool indicating whether the function loss is applicable for the developed TransmissionMode.

The function checks both loss(tm) and loss(tm, t) even if the former is not directly used.

EMGExt.compliance_bidirectional — Function

compliance_bidirectional(tm::TransmissionMode)

Returns a Bool indicating whether the function is_bidirectional is applicable for the developed TransmissionMode.

It is not required if you create a new PipeMode.

EMGExt.compliance_con_rate — Function

compliance_con_rate(tm::TransmissionMode)
compliance_con_rate(tm::PipeMode)

Returns a Bool indicating whether the function consumption_rate is applicable for the developed TransmissionMode.

It is only required if you create a new PipeMode.

The function checks both consumption_rate(tm) and consumption_rate(tm, t) even if the former is not directly used.